7980 revert namex-api config changes and update solr-names-updater config to work with one config file in root, removed business schema from requirements.txt
