+++ services/solr-names-updater/config.py
@@ -1,17 +1,38 @@
-"""Config for initializing the namex-api."""
+# Copyright © 2019 Province of British Columbia
+#
+# Licensed under the Apache License, Version 2.0 (the 'License');
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an 'AS IS' BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""All of the configuration for the service is captured here.
+
+All items are loaded, or have Constants defined here that
+are loaded into the Flask configuration.
+All modules and lookups get their configuration from the
+Flask config, rather than reading environment variables directly
+or by accessing this configuration directly.
+"""
 import os
 import random
 
 from dotenv import find_dotenv, load_dotenv
 
+
 # this will load all the envars from a .env file located in the project root (api)
 load_dotenv(find_dotenv())
 
 CONFIGURATION = {
-    'development': 'namex.config.DevConfig',
-    'testing': 'namex.config.TestConfig',
-    'production': 'namex.config.Config',
-    'default': 'namex.config.Config'
+    'development': 'solr_names_updater.config.DevConfig',
+    'testing': 'solr_names_updater.config.TestConfig',
+    'production': 'solr_names_updater.config.ProdConfig',
+    'default': 'solr_names_updater.config.ProdConfig'
 }
 
 
@@ -21,36 +42,29 @@ def get_named_config(config_name: str = 'production'):
     :raise: KeyError: if an unknown configuration is requested
     """
     if config_name in ['production', 'staging', 'default']:
-        config = Config()
+        app_config = ProdConfig()
     elif config_name == 'testing':
-        config = TestConfig()
+        app_config = TestConfig()
     elif config_name == 'development':
-        config = DevConfig()
+        app_config = DevConfig()
     else:
         raise KeyError(f'Unknown configuration: {config_name}')
-    return config
+    return app_config
 
 
-class Config(object):
-    """Base config (also production config)."""
+class Config():  # pylint: disable=too-few-public-methods
+    """Base class configuration that should set reasonable defaults.
+
+    Used as the base for all the other configurations.
+    """
 
     PROJECT_ROOT = os.path.abspath(os.path.dirname(__file__))
 
     SECRET_KEY = 'a secret'
 
+    SENTRY_DSN = os.getenv('SENTRY_DSN', None)
+
     SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-    NRO_SERVICE_ACCOUNT = os.getenv('NRO_SERVICE_ACCOUNT', 'nro_service_account')
-
-    SOLR_BASE_URL = os.getenv('SOLR_BASE_URL', None)
-    SOLR_SYNONYMS_API_URL = os.getenv('SOLR_SYNONYMS_API_URL', None)
-    NRO_EXTRACTOR_URI = os.getenv('NRO_EXTRACTOR_URI', None)
-    AUTO_ANALYZE_URL = os.getenv('AUTO_ANALYZE_URL', None)
-    AUTO_ANALYZE_CONFIG = os.getenv('AUTO_ANALYZE_CONFIG', None)
-    REPORT_SVC_URL = os.getenv('REPORT_SVC_URL', None)
-    REPORT_TEMPLATE_PATH = os.getenv('REPORT_PATH', 'report-templates')
-
-    ALEMBIC_INI = 'migrations/alembic.ini'
 
     # POSTGRESQL
     DB_USER = os.getenv('DATABASE_USERNAME', '')
@@ -63,15 +77,30 @@ class Config(object):
         password=DB_PASSWORD,
         host=DB_HOST,
         port=int(DB_PORT),
-        name=DB_NAME
+        name=DB_NAME,
     )
-    # ORACLE - LEGACY NRO NAMESDB
-    NRO_USER = os.getenv('NRO_USER', '')
-    NRO_SCHEMA = os.getenv('NRO_SCHEMA', None)
-    NRO_PASSWORD = os.getenv('NRO_PASSWORD', '')
-    NRO_DB_NAME = os.getenv('NRO_DB_NAME', '')
-    NRO_HOST = os.getenv('NRO_HOST', '')
-    NRO_PORT = int(os.getenv('NRO_PORT', '1521'))
+
+    SOLR_FEEDER_API_URL = os.getenv('SOLR_FEEDER_API_URL', None)
+
+    ALEMBIC_INI = 'migrations/alembic.ini'
+
+    NATS_CONNECTION_OPTIONS = {
+        'servers': os.getenv('NATS_SERVERS', 'nats://127.0.0.1:4222').split(','),
+        'name': os.getenv('NATS_CLIENT_NAME', 'namex.solr.names.updater')
+
+    }
+    STAN_CONNECTION_OPTIONS = {
+        'cluster_id': os.getenv('NATS_CLUSTER_ID', 'test-cluster'),
+        'client_id': str(random.SystemRandom().getrandbits(0x58)),
+        'ping_interval': 1,
+        'ping_max_out': 5,
+    }
+
+    SUBSCRIPTION_OPTIONS = {
+        'subject': os.getenv('NATS_SUBJECT', 'namerequest.state'),
+        'queue': os.getenv('NATS_QUEUE', 'namerequest-processor'),
+        'durable_name': os.getenv('NATS_QUEUE', 'namerequest-processor') + '_durable',
+    }
 
     # JWT_OIDC Settings
     JWT_OIDC_WELL_KNOWN_CONFIG = os.getenv('JWT_OIDC_WELL_KNOWN_CONFIG')
@@ -84,37 +113,24 @@ class Config(object):
 
     JWT_OIDC_JWKS_CACHE_TIMEOUT = int(os.getenv('JWT_OIDC_JWKS_CACHE_TIMEOUT', '300'))
 
-    TESTING = False,
+    TESTING = False
     DEBUG = False
 
-    # You can disable NRO updates for Name Requests by setting the variable in your .env / OpenShift configuration
-    DISABLE_NAMEREQUEST_NRO_UPDATES = int(os.getenv('DISABLE_NAMEREQUEST_NRO_UPDATES', 0))
-    DISABLE_NAMEREQUEST_SOLR_UPDATES = int(os.getenv('DISABLE_NAMEREQUEST_SOLR_UPDATES', 0))
-    DISABLE_NAMEREQUEST_NATS_UPDATES = int(os.getenv('DISABLE_NAMEREQUEST_NATS_UPDATES', 0))
-
-    # NATS
-    NATS_SERVERS = os.getenv('NATS_SERVERS', 'nats://localhost:4222')
-    NATS_CLIENT_NAME = os.getenv('NATS_CLIENT_NAME', 'namex.worker')
-    NATS_CLUSTER_ID = os.getenv('NATS_CLUSTER_ID', 'test-cluster')
-    NATS_QUEUE = os.getenv('NATS_QUEUE', 'namerequest-processor')
-    NATS_NR_STATE_SUBJECT = os.getenv('NATS_NR_STATE_SUBJECT', 'namex.event')
-    NATS_EMAILER_SUBJECT = os.getenv('NATS_EMAILER_SUBJECT', 'entity.email')
+    ENVIRONMENT = os.getenv('ENVIRONMENT', 'prod')
 
 
-class DevConfig(Config):
-    """Dev config used for development."""
+class DevConfig(Config):  # pylint: disable=too-few-public-methods
+    """Creates the Development Config object."""
 
-    TESTING = False,
+    TESTING = False
     DEBUG = True
 
-    # We can't run NRO locally unless you're provisioned, you can disable NRO updates for Name Requests by setting the variable in your .env
-    DISABLE_NAMEREQUEST_NRO_UPDATES = int(os.getenv('DISABLE_NAMEREQUEST_NRO_UPDATES', 0))
-    DISABLE_NAMEREQUEST_SOLR_UPDATES = int(os.getenv('DISABLE_NAMEREQUEST_SOLR_UPDATES', 0))
-    DISABLE_NAMEREQUEST_NATS_UPDATES = int(os.getenv('DISABLE_NAMEREQUEST_NATS_UPDATES', 0))
 
+class TestConfig(Config):  # pylint: disable=too-few-public-methods
+    """In support of testing only.
 
-class TestConfig(Config):
-    """Test config used for pytests."""
+    Used by the py.test suite
+    """
 
     DEBUG = True
     TESTING = True
@@ -124,28 +140,24 @@ class TestConfig(Config):
     DB_NAME = os.getenv('DATABASE_TEST_NAME', '')
     DB_HOST = os.getenv('DATABASE_TEST_HOST', '')
     DB_PORT = os.getenv('DATABASE_TEST_PORT', '5432')
-    # Allows for NRO add / update bypass if necessary (for local development)
-    LOCAL_DEV_MODE = os.getenv('LOCAL_DEV_MODE', False)
     # Set this in your .env to debug SQL Alchemy queries (for local development)
+    # pylint: disable=invalid-envvar-default
     SQLALCHEMY_ECHO = 'debug' if os.getenv('DEBUG_SQL_QUERIES', False) else False
     SQLALCHEMY_DATABASE_URI = 'postgresql://{user}:{password}@{host}:{port}/{name}'.format(
         user=DB_USER,
         password=DB_PASSWORD,
         host=DB_HOST,
         port=int(DB_PORT),
-        name=DB_NAME
+        name=DB_NAME,
     )
-
-    # We can't run NRO locally for running our tests
-    DISABLE_NAMEREQUEST_NRO_UPDATES = int(os.getenv('DISABLE_NAMEREQUEST_NRO_UPDATES', 1))
-    DISABLE_NAMEREQUEST_SOLR_UPDATES = int(os.getenv('DISABLE_NAMEREQUEST_SOLR_UPDATES', 0))
-    DISABLE_NAMEREQUEST_NATS_UPDATES = int(os.getenv('DISABLE_NAMEREQUEST_NATS_UPDATES', 1))
 
     # JWT OIDC settings
     # JWT_OIDC_TEST_MODE will set jwt_manager to use
     JWT_OIDC_TEST_MODE = True
-    JWT_OIDC_TEST_AUDIENCE = 'example'
-    JWT_OIDC_TEST_ISSUER = 'https://example.localdomain/auth/realms/example'
+    JWT_OIDC_TEST_AUDIENCE = os.getenv('JWT_OIDC_AUDIENCE')
+    JWT_OIDC_TEST_CLIENT_SECRET = os.getenv('JWT_OIDC_CLIENT_SECRET')
+    JWT_OIDC_TEST_ISSUER = 'https://dev.oidc.gov.bc.ca/auth/realms/fcf0kpqr'
+    # JWT_OIDC_TEST_ISSUER = 'https://sso-dev.pathfinder.gov.bc.ca/auth/realms/sbc'
     JWT_OIDC_TEST_KEYS = {
         'keys': [
             {
@@ -194,3 +206,10 @@ class TestConfig(Config):
     NrQw+2OdQACBJiEHsdZzAkBcsTk7frTH4yGx0VfHxXDPjfTj4wmD6gZIlcIr9lZg
     4H8UZcVFN95vEKxJiLRjAmj6g273pu9kK4ymXNEjWWJn
     -----END RSA PRIVATE KEY-----"""
+
+
+class ProdConfig(Config):  # pylint: disable=too-few-public-methods
+    """Production environment configuration."""
+
+    TESTING = False
+    DEBUG = False
