+++ services/solr-names-updater/requirements.txt
@@ -17,7 +17,6 @@ six==1.15.0
 urllib3==1.26.4
 Werkzeug==1.0.1
 
-git+https://github.com/bcgov/business-schemas.git#egg=registry_schemas
 git+https://github.com/bcgov/namex.git#egg=namex&subdirectory=api
 git+https://github.com/bcgov/namex-synonyms-api-py-client.git#egg=swagger_client
 git+https://github.com/bcgov/namex.git#egg=queue_common&subdirectory=services/common
